🎨 Add validate module
Applye validate module to login.js
