+++ html/login.html
@@ -53,6 +53,6 @@
         </div>
       </div>
     </main>
-    <script src="/js/login.js"></script>
+    <script type="module" src="/js/login.js"></script>
   </body>
 </html>
